pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cmc-csci040-isaiah-bingham-docsum"
version = "0.2.0"
description = "A terminal assistant for chatting with local documents."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "groq",
    "openai",
    "python-dotenv",
    "gitpython",
]

[project.optional-dependencies]
dev = [
    "coverage",
    "flake8",
    "pytest",
    "build",
    "twine",
]

[project.scripts]
chat = "chat:main"

[tool.setuptools]
py-modules = ["chat"]
packages = ["tools"]
